created a separate package for schema and moved user struct to new package
